+++ polymetis/polymetis/python/polymetis/camera_interface.py
@@ -21,7 +21,7 @@ class CameraInterface:
         n_height = image.height
         n_channel = image.channel
         # convert bgr to rgb
-        image_data = np.array(image.image_data).reshape((n_height, n_width, n_channel))
+        image_data = np.array(image.image_data).reshape((n_height, n_width, n_channel)).astype(np.float32)
         image_data[..., :3] = image_data[..., :3][..., ::-1]
         if n_channel == 4:
             camera_metadata = self.stub.GetMetaData(EMPTY)
